backend/requirements.txt
#accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
boto3==1.34.153
botocore==1.34.153
Bottleneck==1.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
datasets==2.20.0
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Jsonpify==1.5.0
Flask-SSE==1.0.0
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.24.0
icecream==2.1.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nanoid==2.0.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
#nvidia-cublas-cu12==12.1.3.1
#nvidia-cuda-cupti-cu12==12.1.105
#nvidia-cuda-nvrtc-cu12==12.1.105
#nvidia-cuda-runtime-cu12==12.1.105
#nvidia-cudnn-cu12==8.9.2.26
#nvidia-cufft-cu12==11.0.2.54
#nvidia-curand-cu12==10.3.2.106
#nvidia-cusolver-cu12==11.4.5.107
#nvidia-cusparse-cu12==12.1.0.106
#nvidia-nccl-cu12==2.20.5
#nvidia-nvjitlink-cu12==12.5.82
#nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
statsmodels==0.14.2
stopwords==1.0.1
sympy==1.13.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.19.1
#torch==2.3.1
#torchaudio==2.3.1
#torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
